Bug 1990248 - Part 1: Add IndexToId with uint64_t argument. r=spidermonkey-reviewers,jandem
Also used in part 2.

Differential Revision: https://phabricator.services.mozilla.com/D265823

+++ js/src/vm/NativeObject.cpp
@@ -2356,7 +2356,7 @@ bool js::NativeGetElement(JSContext* cx, Handle<NativeObject*> obj,
   RootedId id(cx);
 
   if (MOZ_LIKELY(index >= 0)) {
-    if (!IndexToId(cx, index, &id)) {
+    if (!IndexToId(cx, uint32_t(index), &id)) {
       return false;
     }
   } else {
